Move FormBuilder to CreateSchema component
